Move parts of WebViewController to separate files
