+++ Sources/App/WebView/WebViewController.swift
@@ -1041,7 +1041,7 @@ final class WebViewController: UIViewController, WKNavigationDelegate, WKUIDeleg
         SwiftMessages.show(config: config, view: view)
     }
 
-    private func openDebug() {
+    func openDebug() {
         let controller = UIHostingController(rootView: AnyView(
             NavigationView {
                 VStack {
@@ -1069,21 +1069,6 @@ final class WebViewController: UIViewController, WKNavigationDelegate, WKUIDeleg
             }
         ))
         presentOverlayController(controller: controller, animated: true)
-    }
-}
-
-extension String {
-    func matchingStrings(regex: String) -> [[String]] {
-        guard let regex = try? NSRegularExpression(pattern: regex) else { return [] }
-        let nsString = self as NSString
-        let results = regex.matches(in: self, range: NSRange(location: 0, length: nsString.length))
-        return results.map { result in
-            (0 ..< result.numberOfRanges).map {
-                result.range(at: $0).location != NSNotFound
-                    ? nsString.substring(with: result.range(at: $0))
-                    : ""
-            }
-        }
     }
 }
 
@@ -1171,256 +1156,9 @@ extension WebViewController: WKScriptMessageHandler {
 
 extension WebViewController: UIScrollViewDelegate {
     func scrollViewDidScroll(_ scrollView: UIScrollView) {
+        // Prevent scrollView from scrolling past the top or bottom
         if scrollView.contentOffset.y > scrollView.contentSize.height - scrollView.bounds.height {
             scrollView.contentOffset.y = scrollView.contentSize.height - scrollView.bounds.height
         }
-    }
-}
-
-extension ConnectionInfo {
-    mutating func webviewURLComponents() -> URLComponents? {
-        if Current.appConfiguration == .fastlaneSnapshot, prefs.object(forKey: "useDemo") != nil {
-            return URLComponents(string: "https://companion.home-assistant.io/app/ios/demo")!
-        }
-        guard let activeURL = activeURL() else {
-            Current.Log.error("No activeURL available while webviewURLComponents was called")
-            return nil
-        }
-
-        guard var components = URLComponents(url: activeURL, resolvingAgainstBaseURL: true) else {
-            return nil
-        }
-
-        let queryItem = URLQueryItem(name: "external_auth", value: "1")
-        components.queryItems = [queryItem]
-
-        return components
-    }
-
-    mutating func webviewURL() -> URL? {
-        webviewURLComponents()?.url
-    }
-
-    mutating func webviewURL(from raw: String) -> URL? {
-        guard let baseURLComponents = webviewURLComponents(), let baseURL = baseURLComponents.url else {
-            return nil
-        }
-
-        if raw.starts(with: "/") {
-            if let rawComponents = URLComponents(string: raw) {
-                var components = baseURLComponents
-                components.path.append(rawComponents.path)
-                components.fragment = rawComponents.fragment
-
-                if let items = rawComponents.queryItems {
-                    var queryItems = components.queryItems ?? []
-                    queryItems.append(contentsOf: items)
-                    components.queryItems = queryItems
-                }
-
-                return components.url
-            } else {
-                return baseURL.appendingPathComponent(raw)
-            }
-        } else if let url = URL(string: raw), url.baseIsEqual(to: baseURL) {
-            return url
-        } else {
-            return nil
-        }
-    }
-}
-
-extension WebViewController: WebViewControllerProtocol {
-    var overlayedController: UIViewController? {
-        presentedViewController
-    }
-
-    func presentOverlayController(controller: UIViewController, animated: Bool) {
-        DispatchQueue.main.async { [weak self] in
-            self?.dismissOverlayController(animated: false, completion: { [weak self] in
-                self?.present(controller, animated: animated, completion: nil)
-            })
-        }
-    }
-
-    func presentAlertController(controller: UIViewController, animated: Bool) {
-        DispatchQueue.main.async { [weak self] in
-            guard let self else { return }
-            if let overlayedController {
-                overlayedController.present(controller, animated: animated, completion: nil)
-            } else {
-                present(controller, animated: animated, completion: nil)
-            }
-        }
-    }
-
-    func evaluateJavaScript(_ script: String, completion: ((Any?, (any Error)?) -> Void)?) {
-        webView.evaluateJavaScript(script, completionHandler: completion)
-    }
-
-    func dismissOverlayController(animated: Bool, completion: (() -> Void)?) {
-        dismissAllViewControllersAbove(completion: completion)
-    }
-
-    func dismissControllerAboveOverlayController() {
-        overlayedController?.dismissAllViewControllersAbove()
-    }
-
-    func updateSettingsButton(state: String) {
-        // Possible values: connected, disconnected, auth-invalid
-        UIView.animate(withDuration: 1.0, delay: 0, options: .curveEaseInOut, animations: {
-            WebViewAccessoryViews.settingsButton.alpha = state == "connected" ? 0 : 1
-        }, completion: nil)
-    }
-
-    func navigateToPath(path: String) {
-        if let activeURL = server.info.connection.activeURL(), let url = URL(string: activeURL.absoluteString + path) {
-            webView.load(URLRequest(url: url))
-        }
-    }
-
-    func reload() {
-        webView.reload()
-    }
-}
-
-// MARK: - Gestures
-
-extension WebViewController {
-    func handleGestureAction(_ action: HAGestureAction) {
-        switch action {
-        case .showSidebar:
-            webViewExternalMessageHandler.sendExternalBus(message: .init(command: "sidebar/show"))
-        case .backPage:
-            if webView.canGoBack {
-                webView.goBack()
-            }
-        case .nextPage:
-            if webView.canGoForward {
-                webView.goForward()
-            }
-        case .showServersList:
-            Current.sceneManager.webViewWindowControllerPromise.done { controller in
-                controller.selectServer(includeSettings: true).done { server in
-                    if let server {
-                        controller.open(server: server)
-                    }
-                }.catch { error in
-                    Current.Log.error("failed to select server: \(error)")
-                }
-            }
-        case .nextServer:
-            moveToServer(next: true)
-            displayChangeServerHUD(next: true)
-        case .previousServer:
-            moveToServer(next: false)
-            displayChangeServerHUD(next: false)
-        case .showSettings:
-            showSettingsViewController()
-        case .none:
-            /* no-op */
-            break
-        case .openDebug:
-            openDebug()
-        case .searchEntities:
-            showSearchEntities()
-        }
-    }
-
-    private func showSearchEntities() {
-        webView.evaluateJavaScript(WebViewJavascriptCommands.searchEntitiesKeyEvent) { _, error in
-            if let error {
-                Current.Log.error("JavaScript error while trying to open entities search: \(error)")
-            } else {
-                Current.Log.info("Open entities search command sent to webview")
-            }
-        }
-    }
-
-    private func moveToServer(next: Bool) {
-        let servers = Current.servers.all
-        guard servers.count > 1, let currentIndex = servers.firstIndex(of: server) else { return }
-
-        let nextIndex: Int
-        if next {
-            nextIndex = (currentIndex - 1 + servers.count) % servers.count
-        } else {
-            nextIndex = (currentIndex + 1) % servers.count
-        }
-
-        let nextServer = servers[nextIndex]
-
-        Current.sceneManager.webViewWindowControllerPromise.done { controller in
-            controller.open(server: nextServer).done { controller in
-                let hud = MBProgressHUD.showAdded(to: controller.view, animated: true)
-                hud.isUserInteractionEnabled = false
-                hud.mode = .text
-                hud.label.text = nextServer.info.name
-                hud.hide(animated: true, afterDelay: 1.0)
-            }
-        }
-    }
-
-    private func displayChangeServerHUD(next: Bool) {
-        let icon: MaterialDesignIcons = next ? .arrowRightIcon : .arrowLeftIcon
-        let hud = MBProgressHUD.showAdded(to: view, animated: true)
-        hud.isUserInteractionEnabled = false
-        hud.customView = with(IconImageView(frame: CGRect(x: 0, y: 0, width: 37, height: 37))) {
-            $0.iconDrawable = icon
-        }
-        hud.mode = .customView
-        hud.hide(animated: true, afterDelay: 1.0)
-    }
-}
-
-// MARK: - Post onboarding
-
-extension WebViewController {
-    private func postOnboardingNotificationPermission() {
-        // 3 seconds feels a good timin to show this notification after the user has onboarded
-        let delayedSeconds: CGFloat = 3
-        DispatchQueue.main.asyncAfter(deadline: .now() + delayedSeconds) { [weak self] in
-            Task {
-                let notificationCenter = Current.userNotificationCenter
-                let settings = await notificationCenter.notificationSettings()
-                if ![.authorized, .denied].contains(settings.authorizationStatus) {
-                    self?.showNotificationPermissionRequest()
-                }
-            }
-        }
-    }
-
-    private func showNotificationPermissionRequest() {
-        let view = MessageView.viewFromNib(layout: .cardView)
-        var config = SwiftMessages.Config()
-        config.duration = .forever
-        config.presentationStyle = .top
-        config.dimMode = .gray(interactive: true)
-        view.configureContent(
-            title: L10n.PostOnboarding.Permission.Notification.title,
-            body: L10n.PostOnboarding.Permission.Notification.message,
-            iconImage: nil,
-            iconText: nil,
-            buttonImage: MaterialDesignIcons.arrowRightBoldCircleIcon.image(
-                ofSize: .init(width: 35, height: 35),
-                color: Asset.Colors.haPrimary.color
-            ),
-            buttonTitle: nil,
-            buttonTapHandler: { _ in
-                SwiftMessages.hide()
-                UNUserNotificationCenter.current().requestAuthorization(options: .defaultOptions) { _, error in
-                    if let error {
-                        Current.Log.error("Error when requesting notifications permissions: \(error)")
-                    }
-                    DispatchQueue.main.async {
-                        UIApplication.shared.registerForRemoteNotifications()
-                    }
-                }
-            }
-        )
-        view.titleLabel?.numberOfLines = 0
-        view.bodyLabel?.numberOfLines = 0
-
-        SwiftMessages.show(config: config, view: view)
     }
 }
